requirements.txt
Flask==1.1.1
line-bot-sdk==1.16.0
cryptography==2.1.4
chardet==3.0.4
requests==2.23.0
urllib3==1.25.9
flask-bootstrap